nitpicks

+++ src/util/perf_crypto_hash.c
@@ -57,7 +57,7 @@ main (int argc, char *argv[])
   start = GNUNET_TIME_absolute_get ();
   perfHash ();
   printf ("Hash perf took %llu ms\n",
-          GNUNET_TIME_absolute_get_duration (start).value);
+          (unsigned long long) GNUNET_TIME_absolute_get_duration (start).value);
   return 0;
 }
